pyproject.toml
[project]
name = "rag-system"
description = "RAG系统实现，支持文档上传、解析、向量检索和多轮对话"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "langchain-core>=0.2.0",
    "langchain-community>=0.2.0",
    "langgraph>=0.1.0",
    "langsmith>=0.1.0",
    "pypdf>=4.0.0",
    "python-docx>=1.0.0",
    "mistune>=3.0.0",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "accelerate>=0.29.0",
    "httpx>=0.27.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"